+++ src/Motor.h
@@ -41,7 +41,7 @@ class Motor {
     Motor(byte pin1, byte pin2, byte pinSleep, byte encA, byte encB);
     Motor(byte pin1, byte pin2, byte pinSleep, byte encA, byte encB, Constant::Mode mode);
     Motor(byte pin1, byte pin2, byte pinSleep, byte encA, byte encB, Constant::Mode mode, float kp, float kd, float ki);
-    void setSpeed(int speedSetpoint, bool tuning);      
+    void setSpeed(float speedSetpoint, bool tuning);      
     void incrementA();
     void incrementB();
     int getData();
